Resolve performance issues with historical data
I suspected SQL query performance was the issue due to misconfigured indexes,
but turns out that Google App Engine was just hitting the memory cap of 256MB
when processing the query results. I had assumed it auto-scales the memory limit
as needed, but apparently not. The new 512MB memory limit should be sufficient
for now.